+++ src/task.cpp
@@ -2858,7 +2858,7 @@ bool SetConfigTask::Execute()
     document->config = config;
     document->current_code_format->border_color = config.code_block_border_color;
 
-    document->logger->SetLevel((int)config.log_level);
+    document->logger->SetLevel(config.log_level);
 
     if (remake)
         Remake(text->id, false);
